chore: refactor custom service worker's vite build logic
